+++ packages/webview_flutter/webview_flutter/test/webview_controller_test.dart
@@ -516,6 +516,22 @@ void main() {
           .setOnScrollPositionChange(onScrollPositionChange),
     );
   });
+
+  test('setOverScrollMode', () async {
+    final MockPlatformWebViewController mockPlatformWebViewController =
+        MockPlatformWebViewController();
+
+    final WebViewController webViewController = WebViewController.fromPlatform(
+      mockPlatformWebViewController,
+    );
+
+    await webViewController.setOverScrollMode(WebViewOverScrollMode.never);
+    verify(
+      mockPlatformWebViewController.setOverScrollMode(
+        WebViewOverScrollMode.never,
+      ),
+    );
+  });
 }
 
 class TestPlatformWebViewPermissionRequest