+++ packages/webview_flutter/webview_flutter/test/webview_controller_test.mocks.dart
@@ -219,6 +219,7 @@ class MockPlatformWebViewController extends _i1.Mock
   @override
   _i5.Future<void> scrollBy(int? x, int? y) => (super.noSuchMethod(
         Invocation.method(#scrollBy, [x, y]),
+<<<<<<< HEAD
         returnValue: _i5.Future<void>.value(),
         returnValueForMissingStub: _i5.Future<void>.value(),
       ) as _i5.Future<void>);
@@ -235,6 +236,8 @@ class MockPlatformWebViewController extends _i1.Mock
   _i5.Future<void> setHorizontalScrollBarEnabled(bool? enabled) =>
       (super.noSuchMethod(
         Invocation.method(#setHorizontalScrollBarEnabled, [enabled]),
+=======
+>>>>>>> bfb42a62a56ee53010e7c2601cbf1bac993122f2
         returnValue: _i5.Future<void>.value(),
         returnValueForMissingStub: _i5.Future<void>.value(),
       ) as _i5.Future<void>);
@@ -351,6 +354,17 @@ class MockPlatformWebViewController extends _i1.Mock
         Invocation.method(#setOnJavaScriptTextInputDialog, [
           onJavaScriptTextInputDialog,
         ]),
+<<<<<<< HEAD
+=======
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
+
+  @override
+  _i5.Future<void> setOverScrollMode(_i2.WebViewOverScrollMode? mode) =>
+      (super.noSuchMethod(
+        Invocation.method(#setOverScrollMode, [mode]),
+>>>>>>> bfb42a62a56ee53010e7c2601cbf1bac993122f2
         returnValue: _i5.Future<void>.value(),
         returnValueForMissingStub: _i5.Future<void>.value(),
       ) as _i5.Future<void>);