+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.4.0
+
+* Adds support to set the over-scroll mode for the WebView. See `AndroidWebViewController.setOverScrollMode`.
+
+## 4.3.5
+
+* Adds internal wrapper methods for native `WebViewClient`.
+
 ## 4.3.4
 
 * Bumps gradle from 8.0.0 to 8.9.0.